+++ app/Http/Controllers/ServiceStaffController.php
@@ -190,7 +190,6 @@ class ServiceStaffController extends Controller
             'phone' => 'required',
             'whatsapp' => 'required',
             'email' => 'required|email|unique:users,email',
-            'image' => 'required|image|mimes:jpeg,png,jpg,gif|max:2048',
             'password' => 'required|same:confirm-password',
             'commission' => 'required',
             'id_card_front' => 'required',
@@ -235,6 +234,10 @@ class ServiceStaffController extends Controller
             // save the filename to the gallery object and persist it to the database
 
             $input['image'] = $filename;
+        }else{
+            if (file_exists(public_path('staff-images') . '/' . "default.png")) {
+                $input['image'] = "default.png";
+            }
         }
 
         if ($request->youtube_video) {
@@ -246,6 +249,10 @@ class ServiceStaffController extends Controller
                     ]);
                 }
             }
+        }
+
+        if (file_exists(public_path('staff-images') . '/' . "default.png")) {
+            $input['image'] = "default.png";
         }
 
         $staff = Staff::create($input);
@@ -442,7 +449,7 @@ class ServiceStaffController extends Controller
         }
 
         if (isset($request->image)) {
-            if ($staff && $staff->image && file_exists(public_path('staff-images') . '/' . $staff->image)) {
+            if ($staff && $staff->image && $staff->image !== "default.png"  && file_exists(public_path('staff-images') . '/' . $staff->image)) {
                 unlink(public_path('staff-images') . '/' . $staff->image);
             }
 
@@ -601,7 +608,7 @@ class ServiceStaffController extends Controller
      */
     public function destroy(User $serviceStaff)
     {
-        if (isset($serviceStaff->staff->image) && file_exists(public_path('staff-images') . '/' . $serviceStaff->staff->image)) {
+        if (isset($serviceStaff->staff->image) && $serviceStaff->staff->image !== "default.png" && file_exists(public_path('staff-images') . '/' . $serviceStaff->staff->image)) {
             unlink(public_path('staff-images') . '/' . $serviceStaff->staff->image);
         }
